+++ file_processor/meta/changelog.md
@@ -9,6 +9,10 @@
 8. 处理器: 支持 逻辑处理器 (logic) 和 内存处理器 (dram)
 
 # 版本更新
+## 0.0.3-rc3
+1. dram & logic：apc文件处理: 合并行时，数据对比，忽略前后空格 @dukang
+2. dram & logic：apc文件处理: 合并行时，数据对比，N/A 值 替换为空字符串 @dukang
+
 ## 0.0.3-rc2
 1. Issue Fix: T7Code（dram & logic）文件转换为xls时，要使用xlwt引擎，在某些场景下可能文件没法使用 @dukang